requirements.txt
django==4.1.3
djangorestframework==3.14.0
httpie==3.2.1
pyyaml==6.0
uritemplate==4.1.1